Cleanup - move code to util module
